Refactoring of the sound effect engine along the lines of what Saffith started.
* There is now no limit on the number of sound effects that can be used in a quest. You can add new sound effect slots in ZQuest as needed.
* Fixed some cosmetic issues in the ZC sound effect dialog.
* The empty quoted character '' is no longer accepted by the ZScript parser. Previously this was erroneously parsed as the single quote character.

+++ src/parser/BuildVisitors.cpp
@@ -62,7 +62,7 @@ void BuildOpcodes::caseBlock(ASTBlock &host, void *param)
 	c->initCode.clear();
 
 	deallocateRefsUntilCount(startRefCount);
-	while (arrayRefs.size() > startRefCount)
+	while ((int)arrayRefs.size() > startRefCount)
 		arrayRefs.pop_back();
 }
 
@@ -573,7 +573,7 @@ void BuildOpcodes::caseStringConstant(ASTStringConstant& host, void* param)
 
 	// Initialize.
 	c->initCode.push_back(new OSetRegister(new VarArgument(INDEX), new VarArgument(EXP1)));
-	for (int i = 0; i < data.size(); ++i)
+	for (int i = 0; i < (int)data.size(); ++i)
 	{
 		c->initCode.push_back(new OSetImmediate(new VarArgument(INDEX2), new LiteralArgument(i * 10000L)));
 		long value = data[i] * 10000L;
